refactor(lsp): split tsz-lsp request handlers
